refactor!: move step-related classes into namespace
