+++ src/Hamelin/PipelineApplicationBuilder.cs
@@ -1,4 +1,5 @@
 using Hamelin.Extensions;
+using Hamelin.Steps;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
