+++ src/Hamelin/PipelineHost.cs
@@ -1,3 +1,4 @@
+using Hamelin.Steps;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
