+++ tests/Hamelin.Tests.Unit/PipelineHostTests.cs
@@ -1,3 +1,4 @@
+using Hamelin.Steps;
 using Hamelin.Tests.Unit.Helpers;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
